refactor evaluator to be class instance based
