+++ src/utils/Evaluator.js
@@ -1,26 +1,27 @@
 import _ from 'lodash';
 import stringHash from 'string-hash';
-import { JSONLogicEvaluator as CoreEvaluator } from '@formio/core/utils';
+import { Evaluator as CoreEvaluator } from '@formio/core';
 
-export class Evaluator extends CoreEvaluator {
-  static cache = {};
-  static protectedEval = false;
-  static noeval = false;
-  static template(template, hash) {
+export class DefaultEvaluator extends CoreEvaluator {
+  cache = {};
+  protectedEval = false;
+
+  template(template, hash) {
     hash = hash || stringHash(template);
-    if (Evaluator.cache[hash]) {
-      return Evaluator.cache[hash];
+    if (this.cache[hash]) {
+      return this.cache[hash];
     }
     try {
       // Ensure we handle copied templates from the ejs files.
       template = template.replace(/ctx\./g, '');
-      return (Evaluator.cache[hash] = _.template(template, Evaluator.templateSettings));
+      return (this.cache[hash] = _.template(template, DefaultEvaluator.templateSettings));
     }
     catch (err) {
       console.warn('Error while processing template', err, template);
     }
   }
-  static interpolate(rawTemplate, data, _options) {
+
+  interpolate(rawTemplate, data, _options) {
     // Ensure reverse compatability.
     const options = _.isObject(_options) ? _options : { noeval: _options };
     if (typeof rawTemplate === 'function') {
@@ -35,11 +36,11 @@ export class Evaluator extends CoreEvaluator {
 
     rawTemplate = String(rawTemplate);
     let template;
-    if (Evaluator.noeval || options.noeval) {
-      return CoreEvaluator.interpolateString(rawTemplate, data, _options);
+    if (this.noeval || options.noeval) {
+      return this.interpolateString(rawTemplate, data, _options);
     }
     else {
-      template = Evaluator.template(rawTemplate);
+      template = this.template(rawTemplate);
     }
     if (typeof template === 'function') {
       try {
@@ -52,4 +53,15 @@ export class Evaluator extends CoreEvaluator {
     }
     return template;
   }
+}
+
+export let Evaluator = DefaultEvaluator.create();
+
+/**
+ * Set the evaluator to use for evaluating expressions.
+ * @param {Evaluator} override - The new evaluator instance to use.
+ * @returns {void}
+ */
+export function registerEvaluator(override) {
+    Evaluator = override;
 }
